Updated openstack/openstack Project: openstack/keystone  5839ffa23824f503e54beb8ac93d74b192c9d454
Do not require the token_id for converting v3 to v2 tokens

Requiring the token_id as part of the arguments for converting v3 to
v2 tokens results in a more difficult transition to the unification
of the code paths for issuing and validating V2 tokens. This is the
first step in eliminating similar but not-quite the same code paths
for issuing and validating Fernet V2 tokens and all other forms of
V2 tokens.

Change-Id: I348974c8285a1a7b724aee1bfbf2d694d5d77406