Updated openstack/openstack Project: openstack/requirements  58d4e98223727ab6d08df14552e23b306260e4c5
Add pyinotify to global-requirements.txt

This module is needed for oslo.log library.

Is the library actively maintained? -Yes.
https://github.com/seb-m/pyinotify/commits/master.
Is the library good code? -Yes.
Is the library python 3 compatible? -Yes.
Is the library license compatible? -Yes. It uses MIT license.
Is the library already packaged in the distros we target? -No.
Is the function of this library already covered by other libraries
in ``global-requirements.txt``? -No.


License of lib is MIT. It is Python 3 compatible.

Change-Id: Ia7eafdc31438c625716adfcdd17d6af72a361919