Updated openstack/openstack Project: openstack/sahara  ed2c4d54578b473ad609af77cc54de81afd31ff0
Remove test for job type in get_data_sources

There is an explicit test for job type in get_data_sources() with
an early return, but for all job_executons input_id and
output_id will be set to None on creation if they are not present
and data_source_get() will return None when the id is None. So, the
explicit test is not needed -- remove for easier maintenance.

Change-Id: I36f59d9fdb3cde923d09e17af346cff8b518eb8e