Updated openstack/openstack Project: openstack/ironic  4dbf176ad4708e8c802d42e5dfd306f5a0e12ada
Add copyright notice to iRMC driver source code.

This patch adds copyright notice in accordance with "APPENDIX: How to
apply the Apache License to your work" [1].
Also this is a reflection of the gerrit review comment in [2] that is
there's no problem adding a copyright line in the license header in
OpenStack.

[1] http://www.apache.org/licenses/LICENSE-2.0
[2] https://review.openstack.org/#/c/171672/

Change-Id: Ic0d69c735170267640803a9bc707f058306c9da1