Updated openstack/openstack Project: openstack/manila  815866b7d2f22754147a1c9679dcd3f8cd6bf4d3
Fix the log level in scheduler manage

Change the log level 'WARNING' to 'ERROR' in scheduler.manage.py.
Because if the log level is WARNING, The root cause can not be
found quickly when a new share creation fails in scheduler mode.
In other words, this log level is more suitable for 'ERROR'.

Change-Id: Ic6b5a7e84e9dd8c3381b2c3760c9dcc6d434f8ed
Closes-Bug: #1481745