Updated openstack/openstack Project: openstack/tempest  ac6c8cf0b2cd19536e197ed920542c571b64690a
allow non '/' root for dashboard

Currently, there is a hard coded value for the login url for the
dashboard in the tempest config. The root for dashboard needs to
be configurable in tempest as well. This patch allows construction
of the login url from information on the page and the hard-coded
url for the dashboard in the tempest config.

The main goal of this patch is a bridge between current use of '/'
and '/dashboard' for the horizon root in devstack. If desired, post
switch we can update the hardcoded value. And revert this patch.

Related change: I6fbca5cea9e44df160afbccc71bd045437657320

Change-Id: I9a04f936ed6d8c14775a332dc28e903992806c42