Updated openstack/openstack Project: openstack/neutron  70a02288928da523b153ebdf08bd7874f47208f0
Fix gateway port could not retrieve for subnet

In DVR routers when a port is added to a router, then
the command succeeds but the l2 ovs agent raises an
error stating that it could not retrieve the gateway
port for the subnet.

The reason for this is there is mismatch in the
ip_address returned from the subnet for the gateway
port and the actual ip_address of the port that we
added to the router.

Since the subnet info was passed to "get_subnet_for_dvr"
this mismatch was seen.

Instead of passing the subnet we will be passing the
actual fixed_ips with all the details and the subnet
gateway port will be filtered based on the ip_address
in the fixed_ips passed.

Closes-Bug: #1404823

Change-Id: I87a3983951f814350e79f5e2274f4639bb6bc0f5