Updated openstack/openstack Project: openstack/ironic-specs  a4c56b061ffa5746a2d6751440e550064d91aa8a
Update network provider spec

We've agreed informally to use host_id as the flag to decide when ports
should be bound (this is supported and expected by Neutron). Remove the
bind_requested flag from this spec and add detail around the usage of
host_id and the expected behavior.

Change-Id: I52b0861dc3f9fdfd07acf9f0b7a194924cfa9a67
Implements: blueprint network-provider