Updated openstack/openstack Project: openstack/neutron  47f05a6dde698dab653750ffbe886c041cb7d8bc
Disable port creation when invalid MAC address is provided

When a port is manually created with an invalid mac address like
'00:00:00:00:00:00' and associated to neutron router, we see an
exception in L3-agent logs. Since we do not have any valid use-case
to have a port with a mac_address of all zeros, this patch disables
the corresponding port creation/updation request.

Along with all zeros MAC, validation aganist broadcast MAC is also
included in this patch.

DocImpact

Closes-Bug: #1472243
Change-Id: I93875716550dbc1f299aee95c45144e4904af233