Updated openstack/openstack Project: openstack/nova  636425757d77d717f5eb29ce199700afde99011b
Instance destroyed if ironic node in CLEANWAIT

Ironic added a new state 'CLEANWAIT' so nodes that previously had their
provision state as CLEANING will now be in either CLEANING or CLEANWAIT.

This patch updates the ironic driver so that it knows about the CLEANWAIT
state. In particular, when destroying an instance, from nova's perspective,
the instance has been removed if a node is in CLEANWAIT state.

Change-Id: Ib42142d67cf2a640a03ec87cbd5fcc0bfdbeb51f
Closes-Bug: #1479532