Updated openstack/openstack Project: openstack/docs-specs  5f597d997c10563f4b30bfe3f89c0671d931e3b8
Updates the Rework API specification to better surface conceptual

We need to try to make the conceptual information more findable
even while working on a way to integrate these concept topics with
the reference information. So, updating the spec to list the work
items we can do in the meantime.

Change-Id: Id5d3a307e1df9d72939c99c5ffdf2f2aedbba2f5