Updated openstack/openstack Project: openstack/nova  07773000abcee91d06d7e68a9f394a44a3ca001a
Remove tie between system_metadata and extra.flavor

This removes the implicit link between system_metadata and extra.flavor in
the Instance object code when calculating columns to join. This makes us
actually stop querying one when we query for the other, a net win in terms
of database traffic.

This mostly has impacts only on unit tests that mock these calls.

Change-Id: I073305c2fb9bc448c5ed33d0edfab9399bbd319b