Updated openstack/openstack Project: openstack/nova  651bd02b4f875d2fe447461e6a29a067322490c5
Remove last of the plugins/v3 from unit tests

This is part of the v3 cleanup effort. The last remaining files in the
n/t/u/a/o/c/plugins/v3 directory have been moved to the base directory,
and all references to them have been updated to reflect the new paths.

Partial-Bug: #1462901
Change-Id: I38d97803b1a9caea98a1b043b2bc2ad774add3b3