Updated openstack/openstack Project: openstack/glance  d5a8c91d2fc74dfad74d8395970e7c8683b22a57
Fix the db_sync problem in 039 for db2

When using db2 as the database backend, it encounters problem
when running db_sync in step 039 due to "SQL0669N  A system
required index cannot be dropped explicitly", which need to
drop the constraint first then recreate the index. And this
patch also fixes the problem of dropping some indexes which
only are created in non-db2 databases, meanwhile, it will not
break the logic for all other database backends.

Closes-Bug: #1434440

Change-Id: I628bfa9ab1e4e07b71b49fc0e5a6087679b642a6