Updated openstack/openstack Project: openstack/nova  9e27fc1f47c0f1e5b0bdecd29166d59c72d9e903
VMware: map one nova-compute to one VC cluster

This patch changes the VMware VC driver to manage only one compute node
which is a vCenter cluster. The support for multiple nodes was
deprecated in the Kilo release with commit
dcd33757dbcfee3c9a0e623f04a722c91de7ed0c and now it is time to be
completely removed.

DocImpact
The 'cluster_name' property is a single string and multiple values are
no longer accepted.

Change-Id: Ifc17c5049e3ed29c8dd130339207907b00433960