Updated openstack/openstack Project: openstack-infra/project-config  85d5d988c215cbf9e0cc621b8f7a49941759a1b6
Add ubuntu-trusty node to os-ansible-deployment

This patch adds an ubuntu-trusty node-based non-voting
check to os-ansible-deployment to help determine whether
the differing images between hpcloud and raxcloud are
contributing to the gate check failures seen solely in
hpcloud.

The idea is to use the exact same image across both
clouds to see the results.

Change-Id: I7809e3f4b3012586f4aaa0de9d9bc7b0011e720d