Updated openstack/openstack Project: openstack/barbican  8c514a7d1b3ac71de7d4520c8c36a68f9cea32cb
PyKMIP 0.4.0 test requirement

PyKMIP has released version 0.4.0 that is required by the KMIP secret
store. Subsequent patches for the KMIP secret store will be submitted
that utilize the new features.

The 0.4.0 release of PyKMIP includes a new Pie API that is as easy as
pie to use. The KMIP secret store will utilize this functionality.

Change-Id: I568a28288068c6016dc67244d979d3619874a425