Updated openstack/openstack Project: openstack/keystone  59ea878c8c79185d28a3ebf251a63f0d20cb7ce7
Add is_domain field in Project Table

Provides the basic storage of 'is_domain' field in project table.
This flag is not interpreted anywhere yet and is set as False by
default in manager level. We currently do not allow the creation or
seeing a project with this flag set to True via the REST API.
We will allow these operations in subsequent patches.

This attribute is only available in v3. Hence, it's filtered in v2.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Rodrigo Duarte <rodrigods@lsd.ufcg.edu.br>

Change-Id: I190e1dd04b1474703804de1fb212340fe6626c13
Partially-Implements: bp reseller