Updated openstack/openstack Project: openstack/openstack-manuals  8b70fb2e7dfc0e5b5f460626a9bc0bb9eae3bf91
Add links to tables in Config Ref Guide

The Cloud Admin Guide used two common tables that
were auto-generated for the XML version of the book.
These tables are not yet being generated for RST.
This patch replaces the xi:includes for these tables
with links to the tables in the Config Ref Guide.

Change-Id: I36a3c0c8f099ca752aace4c2880585949da3ed4a
Implements: blueprint reorganise-user-guides