Updated openstack/openstack Project: openstack/neutron-fwaas  9130c56fdefbe1aed38aa24fbcb41aa6bb5d8da5
Reduce contention between update and delete

When a delete comes right after an update, the firewall state
can change to ACTIVE right before the RPC message is sent over
the wire, but after the state was transitioned to PENDING_DELETE.

This ends up confusing the agent which sees a firewall being
deleted but with an ACTIVE state.

Deferring the state transition closer to the RPC message minimizes
the chances of a race, and forcing the right state in the payload
ensures that the agent gets what it needs.

Closes-bug: #1483875

Change-Id: I455bbe00395377509beca5ec1d54cf4f7a57152d