Updated openstack/openstack Project: openstack/cinder  64b0f23ac9ea52815095f33b6326436f30bb615d
Update devref unit tests doc

Simple update for our painfully out of date devref docs.

Just a short refresh on unit_tests, there's a lot that could
and probably should be added there, but this change just gets
rid of the obsolete references to Nose and adds info regarding
tox and how to use it.

Change-Id: Ib1ff174ee9ccb8ecb258558d77202062204dfec3