Updated openstack/openstack Project: openstack-infra/project-config  778a6f1125cf48fa255cb829db2fc90b9b21f2e2
Move experimental scenario gate jobs for Sahara to non-voting check

We've tested that jobs works okay and now we'd like to monitor how
this jobs will be working for all change requests. After some
period of time when we'll be sure that jobs works stable enough and
correctly, we'll make them voting (and add to gate pipeline too).

Change-Id: I322a330ff6e458a3de4d4bd5b8584e8f72814950