Updated openstack/openstack Project: openstack/ironic-specs  faf13d86b48e0035d12d74e14d9c579a215ae013
Change "max_retries" to "max_polling_attempts" on OneView driver spec

This patch changes the "max_retries" variable added in ironic.conf to
"max_polling_attempts". The "max_retries" name can imply that the driver
will retry the action, but actually it references to max attempts of
connection with OneView to check if something changed regarding a resource.

Change-Id: If1cc6312b49e1aae4190d8671facfddf787676a1