Updated openstack/openstack Project: openstack/manila  bf3c40e4395c1d06452b2a76de946b1ecf885b3c
glusterfs: volume mapped share layout

The volume management done by gluster_native has
been isolated and captured in a separate layout class.
gluster_native implements only {allow,deny}_access,
for the rest it uses the layout code.

Semantics is preserved with one difference:
the Manila host now is assumed to be set up so that
it can mount the GlusterFS volumes without any
complications. Earlier we assumed to not have cert
based access from Manila host, and turned therefore
SSL off and on on the GlusterFS side. This does not
make sense for the separate, layout-agnostic logic.
(Nb. we already wanted to make the move to set this
assumption, regardless of the layout work.)

Partially implements bp modular-glusterfs-share-layouts

Change-Id: I3cbc55eed0f61fe4808873f78811b6c3fd1c66aa