Updated openstack/openstack Project: openstack/python-keystoneclient  7c1f2995cab9d248fb08221a0a749589e20c266e
Deprecate ServiceCatalog(region_name)

There was a FIXME to deprecate ServiceCatalog's region_name
parameter and property. This is now deprecated.

Note that debtcollector isn't used here since the deprecation
happens on ServiceCatalog's __init__() to catch use in
subclasses of ServiceCatalog. ServiceCatalog also has a
factory function that constructs the correct instance and
the factory function always passes region_name, so it's
always using the deprecated kwarg even when region_name
isn't passed to the factory. It's not worth figuring out
how to do this with debtcollector.

bp deprecations

Change-Id: I0e64712474ca2767f3c0ade919359132450f6776