Updated openstack/openstack Project: openstack/python-glanceclient  17375a422c44641dd459bbde9473e126d874ab82
V2: Do not validate image schema when listing

Previously when listing images via v2, the first image of each batch
was validated against the v2 schema.

This could lead to unpredictable behaviour where a particular image
which failed the schema check may or may not be the first of a batch.

In some cases it also meant that operation by one user could affect the
ability of another other to list images.

Change-Id: I22974a3e3d9cbdd254099780752ae45ff2a557af
Closes-bug: 1477910