Updated openstack/openstack Project: openstack/ceilometer  c6710ff2b5de57d023e4888172fa29fa705cc095
add poll history to avoid duplicate samples

this is a patch until we properly separate polling logic. this
patch adds history so if a resource is defined in separate pipelines
it will only be processed in one. by doing this, we send only a
single copy of each sample and it can be processed appropriately
by the notification agent

this removes duplication check as it is covered by poll history and
removes the pipeline references in polling.

Change-Id: I9c1d3e25740fff5281d796c7bfea2f0e105bb5c5
Closes-Bug: #1480442