Updated openstack/openstack Project: openstack/neutron  4545e3a37abe8ea27313a8898578c9d71faade63
Sync FK constraints in db models with migration scripts

We do have a functional test that compares Neutron's db models with
migration scripts. The comparison is based on alembic library that had a
bug which is gonna be solved in the next release [1]. Once we start
using newer alembic, functional test mentioned above will start failing
due to models and scripts are not in sync.

This patch adds needed constraints discovered by running functional test
locally with dev version of alembic.

Note: There is already a patch [2] that fixes QoS.

[1] https://bitbucket.org/zzzeek/alembic/issues/317
[2] https://review.openstack.org/#/c/214215/

Change-Id: I0d0bddb05f543365d09e592bd81759534de49367
Closes-Bug: 1486936