Updated openstack/openstack Project: openstack/nova  36309829f8f9d09e0c5d92edb51e2ca2a1f5e13a
The API unit tests for serial console use http instead of ws

The serial console will only work if a websocket URL is returned,
e.g. "ws://127.0.0.1:6083/" [1]
The API unit tests use "http" as scheme name, which could lead to
misinterpretation of the capabilities of this feature.They should
use "ws" in the tests to avoid misleading.

[1] table 3.53;
http://docs.openstack.org/kilo/config-reference/content/list-of-compute-config-options.html

Change-Id: I9183009398d5f4f2bcac6f78e6347b8923c6404e
Closes-Bug: #1459943