Updated openstack/openstack Project: openstack/neutron-lbaas  9e71f969b30860dbb6a7e017f867d11f2607c984
Prevent deletion of a subnet with lbaas v1 pool

When using LBaaS and trying to delete a subnet, neutron has no way of
knowing if the subnet is associated to some pool. As a result, the
subnet is deleted but the pool remains associated to the (now
nonexistent) subnet_id.  This patch adds a check in LBaaS side to
prevent such a case, using the callbacks system.

Closes-Bug: #1413817
Change-Id: I1a8893453b0b623df792fdabcf540163007a3f33
Depends-on: I3d5e231b67c72ffd919c92d65b57da56c63e053c