Updated openstack/openstack Project: openstack/keystone  c720a066b789bace591975a10b3b99739d344f28
Enhance tests for saml2 signing exception logging

There was no test that verified that the log message was generated
correctly when there was an exception from the xmlsec1 binary. The
test shows that the log message isn't generated correctly but
instead causes a TypeError to be raised.

The test__sign_assertion_fileutils_exc test is also enhanced to
verify that the error is logged correctly.

Change-Id: I5b7c22d27bad241339fdacae8b61973d45d55d30
Partial-Bug: 1484735