Updated openstack/openstack Project: openstack-infra/puppet-zanata  1a9214cc66da34fe1ebbd4d403ad280738b7d87d
Switch to a class require for deployment

Currently the second requirement for the ROOT.war deployment is an
Exec requirement that comes deep from within biemond-wildfly. This
is a layering violation and makes it impossible to upgrade the module.
Require the install class itself, rather than a part of its internals.

Change-Id: Ie381fcddd98a65c019a44cbde9054db6fafb88f4