Updated openstack/openstack Project: openstack/neutron  cedbc942c5b0b99f11d7b5727ecd47c144b5994a
fullstack: use migration scripts to create db schema

Previously, we used create_all() based on models. We don't use
create_all() in production code and there is no guarantee models and
scripts are in sync even though we have a good functional test that
validates that. There are still pieces that can't be compared by
alembic.

Change-Id: I72fa67811f0763298416e6e084a8b9b86619795b
Closes-Bug: 1486528