Updated openstack/openstack Project: openstack/cinder  45fb47c724724943b60e4e5c09c50b151f2f5350
Parse out SolidFire account from api response

In the SolidFire create-cache account routine we broke the response
check and are trying to use the entire api response as if it were an
account object.

This patch just pulls the actual account object out of the response
instead of using the entire API response.

Change-Id: Ief2d9ae494475f5ba3688c521e11d51a244b5c28
Closes-Bug: #1486744