Updated openstack/openstack Project: openstack/nova  89199bf2cb87a81a145cdba4af89ba1a9d78d9b8
don't report service group connection events as errors in dbdriver

During the upgrade of the API surface with running computes we're
going to see disconnects and reconnects for the service group API (as
conductor is gone). This is not an ERROR, but a normally expected
thing.

Change the connect to an INFO

Change the disconnect to a WARNING (if it happens a lot someone should
look into it).

In both cases try to be a lot more specific, because 'model server' is
really vague and not a term we use in ops facing documentation.

Change-Id: I55417a5b91282c69432bb2ab64441c5cea474d31