Updated openstack/openstack Project: openstack/neutron  5282e80941050c17c8286011d8a954137bfa612c
qos: Delete bw limit rule when policy is deleted

We need to add ON DELETE CASCADE to qos_policy_id on bw limit rule table
in order to delete policy successfully. There is a migration script that
creates db scheme with correct foreign key constraint but we miss this in
models. Currently, we have a functional test that guarantees parity
between migration scripts and models but we don't have guaranteed foreign
keys parity due to alembic bug [1].

https://bitbucket.org/zzzeek/alembic/issues/317

Change-Id: I06fa32dd11a5a52a80ae5a7952f8b32511c3f39d
Closes-Bug: 1485926