Updated openstack/openstack Project: openstack/cinder  5439237455a51b931eefe2ac42367802145bac5d
Add SolidFire svip to config options

The SolidFire driver uses a simple method of gathering
the Storage Virtual IP, it just queries the Cluster and
asks for the SVIP. This is great almost all the time,
BUT it doesn't take into consideration that the deployer
may be using VLAN's and accessing the cluster from multiple
clouds, or multiple networks.

This patch just adds a config option to specify an SVIP
if the admin has configured VLANs in their environment and
on their cluster.  The default behavior is to still use the
global SVIP.

Change-Id: I10b784f903cd8d16580110f01d338a3ad867eada
Closes-Bug: #1487194