Updated openstack/openstack Project: openstack-infra/nodepool  6b80ca4e20468ca253397121e7a2b954c156c360
Move nodepool test checks into nodepool

It doesn't scale to have to change project-config every time we need to
tweak the post devstack run check of nodepool. Instead put a script in
nodepool that can be run by the job to check up on the service once it
is deployed.

Change-Id: Ie50c07b0581162c6fa297c0e1a5e38b4d1a7ca40