Updated openstack/openstack Project: openstack-infra/project-config  2b04611ebff79f81b4d339eda30fa63867db10a6
Run the post nodepool deploy checks from nodepool

Move the checks out of the nodepool dsvm job into nodepool itself so
that we can update the checks as we make changes to nodepool. This
simplifies making the tests more robust over time as we won't have to
update project-config too and it will be self testing.

Depends-on: Ie50c07b0581162c6fa297c0e1a5e38b4d1a7ca40
Change-Id: I66f22e73676000dc6fafc84baa816475aca1a32a