Updated openstack/openstack Project: openstack/nova  0133eb4923777f47b790845cc8d4369400ca8fbc
Handle MessageTimeout to MigrationPreCheckError

There are a few checks before live-migration, if any of the
check failed, the live-migration can't be done. However
during _call_livem_checks_on_host check, because it's a
RPC call ,so it might result in a MessageTimeout exception.
If this occurs, we should safely revert its state,
since no real opperation occured.

This patch translates the MessageTimeout to
MigrationPreCheckError and leverage existing MigrationPreCheckError
processing to revert instance to normal state instead of error

Closes-Bug: 1435633

Change-Id: I8b484abb6650e14e2d225ca5e476d1fa7a6ee990