Updated openstack/openstack Project: openstack-infra/storyboard-webclient  9226d4c96016e10ca1e1821c69192125528baef7
Allow scroll when navbar content exceeds height of screen

The navbar currently hides all content that overflows the space. This
means that some icons do not display for admin users, and users at
various zooms. This is a short-term solution; ideally I'd like the
navigation bar to resize appropriately-- but this is a simpler fix for
now.

Change-Id: Iea1cc310baa75308d1063adcd0814cb99b5f13f7