Updated openstack/openstack Project: openstack-infra/elastic-recheck  3121cb4c7bd991bcfe3df34cf51a03ae52087448
Add query for server GET 500

7 hits in the last 48 hours, but since we've lost the last week
of logstash data this is probably worse.

This is a race when listing servers for all tenants where we
get the initial list, but when lazy-loading attributes on the instances,
like flavor, the instances are deleted (from other tests running in
parallel).

We can't use the InstanceNotFound in the n-api logs because that only
results in 25% failure rates.

Change-Id: I6c4588985548da3d5af2f4b3f76811324efad7ae
Related-Bug: #1487570